Style contact page mobile view
