+++ src/css/contact.module.css
@@ -1,0 +1,106 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: flex-start;
+    width: 90%;
+    margin: auto;
+    padding: 64px 0 32px 0;
+  }
+
+.container h2 {
+  font-weight: 600;
+  font-size: 36px;
+  line-height: 44px;
+  letter-spacing: -0.02em;
+  color: #101828;
+  margin-bottom: 16px;
+}
+
+.container p {
+  font-weight: 400;
+  font-size: 18px;
+  line-height: 28px;
+  color: #475467;
+}
+
+.form {
+  width: 100%;
+  margin-top: 48px;
+}
+
+.inputContainer {
+  display: flex;
+  justify-content: left;
+  flex-direction: column;
+  align-items: left;
+  gap: 6px;
+  margin-bottom: 24px;
+}
+
+.inputContainer label {
+font-weight: 500;
+color: #344054;
+}
+
+.inputContainer input {
+  width: 100%;
+  padding: 14px;
+  background: #FFFFFF;
+  border: 1px solid #D0D5DD;
+  box-shadow: 0px 1px 2px rgba(16, 24, 40, 0.05);
+  border-radius: 8px;
+}
+
+input::placeholder {
+  font-weight: 400;
+  font-size: 16px;
+  color: #667085;
+}
+
+.textarea {
+padding: 12px 14px;
+width: 100%;
+height: 132px;
+background: #FFFFFF;
+border: 1px solid #D0D5DD;
+box-shadow: 0px 1px 2px rgba(16, 24, 40, 0.05);
+border-radius: 8px;
+font-size: 16px;
+line-height: 24px;
+color: #667085;
+}
+
+.conditions {
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+  gap: 12px;
+  margin-top: 26px;
+  margin-bottom: 32px;
+}
+
+.checkbox {
+  background: #FFFFFF;
+  border: 1px solid #D0D5DD;
+  border-radius: 6px;
+}
+
+.conditions span {
+  font-weight: 400;
+  font-size: 16px;
+  color: #475467;
+}
+
+.submitButton {
+padding: 12px 20px;
+background: #1570EF;
+border: 1px solid #1570EF;
+box-shadow: 0px 1px 2px rgba(16, 24, 40, 0.05);
+border-radius: 8px;
+font-weight: 600;
+font-size: 16px;
+margin-bottom: 64px;
+color: #FFFFFF;
+width: 100%;
+}
