--- src/pages/Contact.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Contact = () => (
-  <div>
-    contact
-  </div>
-);
-
-export default Contact;
